PR feedback fixup
