+++ stream_alert/shared/lookup_tables/errors.py
@@ -1,15 +1,15 @@
 
 class LookupTablesError(RuntimeError):
-    pass
+    """Generic class for errors raised from LookupTables systems"""
 
 
 class LookupTablesInitializationError(LookupTablesError):
-    pass
+    """Any error raised when a specific table/driver is attempting to initialize"""
 
 
 class LookupTablesCommitError(LookupTablesError):
-    pass
+    """Any error raised when a LookupTable or driver fails to successfully commit changes"""
 
 
 class LookupTablesConfigurationError(LookupTablesError):
-    pass
+    """Errors raised that detect a misconfiguration for any LookupTables system"""
