+++ tests/unit/stream_alert_shared/lookup_tables/test_driver_s3.py
@@ -23,7 +23,7 @@ from moto import mock_s3
 from nose.tools import assert_equal, assert_raises
 
 from stream_alert.shared.config import load_config
-from stream_alert.shared.lookup_tables.drivers import construct_persistence_driver
+from stream_alert.shared.lookup_tables.drivers_factory import construct_persistence_driver
 from stream_alert.shared.lookup_tables.errors import LookupTablesInitializationError
 from tests.unit.helpers.aws_mocks import put_mock_s3_object
