+++ tests/unit/stream_alert_shared/lookup_tables/test_table.py
@@ -15,7 +15,7 @@ limitations under the License.
 """
 from nose.tools import assert_equal
 
-from stream_alert.shared.lookup_tables.drivers import construct_persistence_driver
+from stream_alert.shared.lookup_tables.drivers_factory import construct_persistence_driver
 from stream_alert.shared.lookup_tables.table import LookupTable
